% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_groupwise.R
\name{cfa_groupwise}
\alias{cfa_groupwise}
\title{Model Selection: cfa_groupwise}
\usage{
cfa_groupwise(model = NULL, data, group, items = NULL, ordered = F)
}
\arguments{
\item{model}{explicit lavaan model. use lavaan syntax. If model is not explictly specified, the function will run a CFA assuming all columns belong to one latent variable.}

\item{data}{dataframe. a multilevel dataset}

\item{group}{character. required. the variable uses for series of CFA.}

\item{items}{vector or quos(). default to NULL. create a lavaan formula if the formula is not explicitly set in model. if you want to use tidyselect syntax, wrap it in quos() (e.g. quos(contains('Q'))). This argument will be ignored if model is specified}

\item{ordered}{logical. default is `F`. If it is set to `T``, lavaan will treat it as a ordinal variable and use DWLS instead of ML}
}
\value{
return a dataframe with cfi, rmsea, tli for each group
}
\description{
This function will run a series of CFA (n = length(group)) with respect to each group. The function is intended to help you get a better understanding of which group has abnormal fit indicator
}
\examples{

}
