% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_items_optimize.R
\name{cfa_items_optimize}
\alias{cfa_items_optimize}
\title{Model Selection: CFA Items Optimizer}
\usage{
cfa_items_optimize(
  data,
  items,
  threshold = 0.9,
  optimize_indicator = "cfi",
  start_columns = NULL,
  group = NULL,
  ordered = F
)
}
\arguments{
\item{data}{dataframe}

\item{items}{vector or quos(). required. create a lavaan formula. if you want to use tidyselect syntax, wrap it in quos() (e.g. quos(contains('Q'))).}

\item{threshold}{the CFI threshold for the optimizer to stop. Recommend setting to 0.95 or 0.9 (the recommended cutoff based on the literature)}

\item{optimize_indicator}{the optimize_indicator. default to CFI. You can probably change to rmsea or tli but you must be aware of the cutoff. Haven't tested other indicator except CFI}

\item{start_columns}{default function or vector of length 3 or greater. default as NULL. CFA required at least 3 observed factors. By default, it will use the cfa_model_optimize_init_factors to find the best 4 initial factors.}

\item{group}{optional character. the nested variable for multilevel dataset (e.g., Country).  If it is specified, it may requires a very long computation time. You can specified the start_columns to avoid lengthy time of optimizing the initial 4 factors.}

\item{ordered}{logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML}
}
\value{
return a vector of the optimized observed factors
}
\description{
The function help you to select the best item for CFA using a data-driven approach. In short, it computed a series of CFA using a stepwise apporach, then it select and retain the best observed factor. Then, it compute again for the best observed factor until the CFI is below the set threshold.
}
