% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_items_stepwise.R
\name{cfa_items_stepwise}
\alias{cfa_items_stepwise}
\title{Model Selection: Stepwise CFA by items}
\usage{
cfa_items_stepwise(data, items, group = NULL, ordered = F)
}
\arguments{
\item{data}{dataframe}

\item{items}{vector or quos(). default to NULL. create a lavaan formula. if you want to use tidyselect syntax, wrap it in quos() (e.g. quos(contains('Q'))).}

\item{group}{optional character. default is NULL. the nested variable for multilevel dataset (e.g., Country)}

\item{ordered}{logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML}
}
\value{

}
\description{
The model will run a series of CFA (similiar to stepwise regression) by adding observed factor to the CFA 1 by 1. Currently, it only support stepwise CFA for one latent factor CFA.
}
\examples{

}
