% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_metric_summary.R
\name{cfa_metric_summary}
\alias{cfa_metric_summary}
\title{Confirmatory factor analysis metric model}
\usage{
cfa_metric_summary(
  model = NULL,
  data,
  group,
  items = NULL,
  summary_item = c("cfi", "rmsea", "tli"),
  ordered = F,
  return_result = "model",
  quite = F,
  group_partial = NULL
)
}
\arguments{
\item{model}{explicit lavaan model. use lavaan syntax. If model is not explictly specified, the function will run a CFA assuming all columns belong to one latent variable.}

\item{data}{dataframe}

\item{group}{character. required. the nested variable for multilevel dataset (e.g., Country)}

\item{items}{vector or quos(). default to NULL. create a lavaan formula if the formula is not explicitly set in model. if you want to use tidyselect syntax, wrap it in quos() (e.g. quos(contains('Q'))). This argument will be ignored if model is specified}

\item{summary_item}{vector. default is cfi, rmsea, and tli. See lavaan for more option}

\item{ordered}{logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML}

\item{return_result}{character. default is 'model'. You can set it to 'summary' for fit measures}

\item{quite}{default as F. If set to true, it will not print the statement that tells you what the lavaan model is going to run.}

\item{group_partial}{items for partial equivalence. The form should be c('DV =~ item1', 'DV =~ item2').}
}
\value{

}
\description{
The function is a wrapper of the lavaan::cfa. By default, it fits a lavaan model based on the data by treating everything columns measure a single latent variable.
}
\examples{


}
