% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_model_optimize_init_factors.R
\name{cfa_model_optimize_init_factors}
\alias{cfa_model_optimize_init_factors}
\title{Model Selection: CFA Initial Items Optimizer}
\usage{
cfa_model_optimize_init_factors(
  data,
  items,
  group = NULL,
  ordered = F,
  fac_num = 4
)
}
\arguments{
\item{data}{dataframe}

\item{items}{vector or quos(). required. create a lavaan formula. if you want to use tidyselect syntax, wrap it in quos() (e.g. quos(contains('Q'))).}

\item{group}{optional character. the nested variable for multilevel dataset (e.g., Country). If it is specified, it may requires a very long computation time.}

\item{ordered}{logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML}

\item{fac_num}{numeric. number of initial factors to be selected. default to 4. It is the minimum number for overidentified (non-saturated) model, so the model can produce a CFI value.}
}
\value{

}
\description{
This function uses a for loop to find the CFI for all possible combinations of the factors.
}
