% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_invariance.R
\name{measurement_invariance}
\alias{measurement_invariance}
\title{Measurement Invariance}
\usage{
measurement_invariance(
  model = NULL,
  data,
  items = NULL,
  group,
  ordered = F,
  group_partial = NULL
)
}
\arguments{
\item{model}{explicit lavaan model. use lavaan syntax. If model is not explictly specified, the function will run a CFA assuming all columns belong to one latent variable.}

\item{data}{dataframe}

\item{items}{vector or quos(). default to NULL. create a lavaan formula if the formula is not explicitly set in model. if you want to use tidyselect syntax, wrap it in quos() (e.g. quos(contains('Q')))}

\item{group}{character. required for metric model.}

\item{ordered}{logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML}
}
\value{
return the compareFit object
}
\description{
This function uses the semTools::compareFit. For now, it will only compute the test measures between the configural and metric model
}
\examples{

}
