% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_invariance_optimize.R
\name{measurement_invariance_optimize}
\alias{measurement_invariance_optimize}
\title{Model Selection: Measurement Invariance Optimizer}
\usage{
measurement_invariance_optimize(
  data,
  items,
  threshold = 0.001,
  group,
  ordered = F
)
}
\arguments{
\item{data}{dataframe}

\item{items}{vector or quos(). required. create a lavaan formula. if you want to use tidyselect syntax, wrap it in quos() (e.g. quos(contains('Q'))).}

\item{threshold}{the delta_CFI threshold. Default is 0.001}

\item{group}{required character. the nested variable for multilevel dataset (e.g., Country).}

\item{ordered}{logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML}
}
\value{
return a lavaan syntax model
}
\description{
The function help you to dplyr::select the items for that is fit with measurement invariance using a data-driven approach. It will take very long since it needs to compute the configural and the metric CFA many many times.
}
\examples{

}
