% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_Plot.R
\name{SEM_Plot}
\alias{SEM_Plot}
\title{the function is not yet ready}
\usage{
SEM_Plot(model)
}
\arguments{
\item{model}{a model object that is acceptable in semPlot::semPaths}
}
\value{

}
\description{
the function is not yet ready
}
\examples{

}
